StorageKey was not serialized before
Fixed serialization of custom types
GetVariable<char> won't work because if not explicitly told, Json will be converted to a string. This should be hopefully the only exception
The VariableStorage Json is split into chunks. The game's save system has a short.MaxValue limit in one entry size. The mitigate it by splitting the entry into an array