Added simplified flags to MBObjectVariableStorage interface
These are a convenience construct: either a flag is set, or it
isn't. Implemented with a separate namespace in another
ConcurrentDictionary, except the dictionary value type is bool.
A uint might be more compact in serialized form if numbers don't
get quoted ("<storagekey>" => 1), but I forget if they do. I
wanted to use a ConcurrentBag, but apparently that's a no-go for
JsonConvert.