FIX MBObjectExtensionDataStore: ensure non-null containers on load
Under some strange circumstances, it appears that SyncData<T> will
set its ref argument to null (and cast it back to T) when loading.
This was leading to the dictionaries in MBObjectExtensionDataStore
being null when it came time to save, violating the nullable API
invariant of ReleaseOrphanedEntries<T>() and causing a null ref
exception.

This fix corresponds to the following crash report submitted via
NexusMods by spartasmaximus ("Error in 1.5.4" original title):
https://drive.google.com/file/d/1ELBRfUHZz3KyjNOq3IyptMxdiy5g5Xa_/view?usp=sharing